Added widget which can host a tool bar. Added widget to wrap an EiffelVision2 widget. Added widget to wrap and EiffelVision2 widget for a specific window. Added an editor based widget. EV widgets are only destroy if a widget was not added to a window. Preference widgets are no longer a wrapped widget. Renamed validation widget to be more consistent. Renamed restricted formatter widget to be more consistent.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@76956 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/Eiffel/interface/graphical/foundations/widgets/es_widget.e
@@ -49,7 +49,8 @@ feature {NONE} -- Clean up
 	internal_recycle
 			-- <Precursor>
 		do
-			if is_initialized then
+			if is_initialized and not widget.has_parent then
+					-- Widget does not have a parent so it needs to be destroyed explicitly.
 				widget.destroy
 			end
 			Precursor {ES_TOOL_FOUNDATIONS}
@@ -257,7 +258,7 @@ feature {NONE} -- Factory
 		end
 
 ;note
-	copyright:	"Copyright (c) 1984-2008, Eiffel Software"
+	copyright:	"Copyright (c) 1984-2009, Eiffel Software"
 	license:	"GPL version 2 (see http://www.eiffel.com/licensing/gpl.txt)"
 	licensing_options:	"http://www.eiffel.com/licensing"
 	copying: "[
@@ -270,22 +271,22 @@ feature {NONE} -- Factory
 			(available at the URL listed under "license" above).
 			
 			Eiffel Software's Eiffel Development Environment is
-			distributed in the hope that it will be useful,	but
+			distributed in the hope that it will be useful, but
 			WITHOUT ANY WARRANTY; without even the implied warranty
 			of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-			See the	GNU General Public License for more details.
+			See the GNU General Public License for more details.
 			
 			You should have received a copy of the GNU General Public
 			License along with Eiffel Software's Eiffel Development
 			Environment; if not, write to the Free Software Foundation,
-			Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA
+			Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
 		]"
 	source: "[
-			 Eiffel Software
-			 356 Storke Road, Goleta, CA 93117 USA
-			 Telephone 805-685-1006, Fax 805-685-6869
-			 Website http://www.eiffel.com
-			 Customer support http://support.eiffel.com
+			Eiffel Software
+			5949 Hollister Ave., Goleta, CA 93117 USA
+			Telephone 805-685-1006, Fax 805-685-6869
+			Website http://www.eiffel.com
+			Customer support http://support.eiffel.com
 		]"
 
 end
